requirements.txt
{# The list of all python modules we need #}

django==1.8.2
djangorestframework==3.1.3
django-extensions==1.5.5
djangorestframework-jwt==1.6.0
Werkzeug==0.10.4
sniffer==0.3.5
